HSWatch Connection Module Added
It was implemented the HSWatch Connection Module, the module that will be used in the Workshops and/or as dependency for projects similar of this one.

The idea behind it was to make the connection part (the most complex one) as modular as possible within the project context and to make this section as independent as possible so it could be used in other projects more freely besides this one.