Android Manifest updated to the current compileSDK version
Since there were some changes in the previous commit in respect of the dependencies version, the manifest was necessary to change to adjust those changes, which was not committed up until now.
Hopefully, these changes solve this problem and the following commits will be just the advance of the module.