HSWatch connection submodule implemented
The submodule for the bluetooth connection is now complete and ready for tests. Therefore, this submodule will be, ideally, a gitsubmodule of a repository, which will make, possibly, in the future commits some modifications to the main repositories

As for the main changes in this commit. This one is just a collective addition and copying of existing files from the bluetooth connection - such as MainService (now HSWService in the submodule) or the threads (HSWThreadConnection for instance and in the same manner of the HSWService) - into its own submodules. Not only that, but also these files were modified to be more modular than the previous ones, so it's possible to implement them in different applications aside from the HSWatch.